+++ lib/alga_view/all_apps/alga_app_view.dart
@@ -1,9 +1,6 @@
 import 'package:alga/constants/import_helper.dart';
 import 'package:alga/models/app_atom.dart';
 import 'package:alga/models/app_category.dart';
-import 'package:alga/utils/hive_boxes/favorite_box.dart';
-import 'package:auto_size_text/auto_size_text.dart';
-import 'package:go_router/go_router.dart';
 
 import 'alga_app_item.dart';