+++ lib/routers/app_routes.dart
@@ -29,7 +29,7 @@ import 'package:alga/tools/js_tools/quick_js_tool/quick_js_view.dart';
 import 'package:alga/tools/server_tools/static_server_tool/static_server_tool_view.dart';
 import 'package:alga/tools/text_tools/date_parser/date_parser_view.dart';
 import 'package:alga/tools/text_tools/markdown_preview/markdown_preview_view.dart';
-import 'package:alga/tools/text_tools/regex_tester/regex_tester_view.dart';
+import 'package:alga/tools/text_tools/regex_tester/regex_tester_app_view.dart';
 import 'package:alga/views/favorite_view.dart';
 import 'package:alga/views/settings_view.dart';
 import 'package:go_router/go_router.dart';
@@ -91,7 +91,7 @@ final appRouter = GoRouter(
             ),
             GoRoute(
               path: AppAtom.regex.path,
-              builder: (context, state) => const RegexTesterView(),
+              builder: (context, state) => const RegexTestAppView(),
             ),
             GoRoute(
               path: AppAtom.markdown.path,