+++ lib/tools/text_tools/regex_tester/regex_tester_text_builder.dart
@@ -2,21 +2,25 @@ import 'package:flutter/material.dart';
 
 import 'package:extended_text_field/extended_text_field.dart';
 
-import 'package:alga/tools/text_tools/regex_tester/regex_tester_provider.dart';
-
 class RegexTesterTextBuilder extends SpecialTextSpanBuilder {
-  final RegexTesterProvider provider;
-  RegexTesterTextBuilder(this.provider);
+  RegexTesterTextBuilder(
+    this.reg, {
+    required this.primaryColor,
+    required this.onPrimary,
+  });
+  final RegExp? reg;
+  final Color primaryColor;
+  final Color onPrimary;
   @override
   TextSpan build(String data,
       {TextStyle? textStyle, SpecialTextGestureTapCallback? onTap}) {
     if (data.isEmpty) return const TextSpan(text: '');
-    if (provider.reg == null) return TextSpan(text: data, style: textStyle);
+    if (reg == null) return TextSpan(text: data, style: textStyle);
 
     String cacheData = data;
     final spans = <InlineSpan>[];
 
-    RegExpMatch? currentMatch = provider.reg!.firstMatch(cacheData);
+    RegExpMatch? currentMatch = reg!.firstMatch(cacheData);
 
     while (currentMatch != null) {
       spans.add(TextSpan(text: cacheData.substring(0, currentMatch.start)));
@@ -25,19 +29,20 @@ class RegexTesterTextBuilder extends SpecialTextSpanBuilder {
           currentMatch.start,
           currentMatch.end,
         ),
-        style: const TextStyle(backgroundColor: Colors.blue),
+        style: TextStyle(
+          backgroundColor: primaryColor,
+          color: onPrimary,
+        ),
       ));
       if (currentMatch.end <= cacheData.length) {
         cacheData = cacheData.substring(currentMatch.end);
       }
-      currentMatch = provider.reg!.firstMatch(cacheData);
+      currentMatch = reg!.firstMatch(cacheData);
     }
     spans.add(TextSpan(text: cacheData));
 
-    return TextSpan(children: spans);
+    return TextSpan(children: spans, style: textStyle);
   }
-
-  matched() {}
 
   @override
   SpecialText? createSpecialText(String flag,